Fixed nspikes indexing
I'd used the full ii.getindex for some of the indexing operations,
forgetting that nspikes does not have a cond axis.
This did not have an actual impact, however, since the cond index
returned by getindex(no cond here) is the full slice.
Still, it's good to observe the correct way of doing this...