Update inhibitory model (see long form)
This was meant to be a preparatory change to
simply align the inh model with the exc for vsyn
recording, which I'd missed (oops).
In the process, I noticed that inh du/dt used
g_exc instead of g_exc_nox, i.e., was actually
synonymous with dv/dt (oops).
The upshot is that all analysis relying on u
(i.e., all analysis referring to Depression) was
flawed. The bare run data, however (spikes etc)
are unaffected.